Fixed timelines +-1 and import errors
