+++ sampo/pipeline/__init__.py
@@ -1,8 +1,17 @@
 from enum import Enum
 
+class PipelineError(Exception):
+    """
+    Raised when any pipeline error occurred.
+
+    This is a kind of 'IllegalStateException', e.g. raising this
+    indicates that the corresponding pipeline come to incorrect internal state.
+    """
+    def __init__(self, message: str):
+        super().__init__(message)
+
 from sampo.pipeline.base import InputPipeline
 from sampo.pipeline.default import DefaultInputPipeline
-from sampo.pipeline.exception import PipelineError
 
 
 class PipelineType(Enum):
