+++ sampo/scheduler/timeline/just_in_time_timeline.py
@@ -103,6 +103,7 @@ class JustInTimeTimeline(Timeline):
             working_time = work_estimator.estimate_time(dep_node.work_unit, worker_team)
             new_finish_time = dep_st + working_time
 
+        new_finish_time -= 1
         exec_time = new_finish_time - c_st
 
         max_material_time = self._material_timeline.find_min_material_time(node.id, c_st,
@@ -285,9 +286,7 @@ class JustInTimeTimeline(Timeline):
             else:
                 lag, working_time = 0, work_estimator.estimate_time(node.work_unit, workers)
             c_st = max(c_ft + lag, max_parent_time)
-            # +1 because we think that work ends in the rest of the time unit
-            if working_time > 0:
-                working_time -= 1
+
             new_finish_time = c_st + working_time
 
             deliveries, _, new_finish_time = self._material_timeline.deliver_materials(dep_node.id, c_st,
